+++ client/src/game/api/emits/shape/options.ts
@@ -16,6 +16,7 @@ export const sendShapeSetIsToken = sendSimpleShapeOption<boolean>("Shape.Options
 export const sendShapeSetBlocksMovement = sendSimpleShapeOption<boolean>("Shape.Options.MovementBlock.Set");
 export const sendShapeSetBlocksVision = sendSimpleShapeOption<boolean>("Shape.Options.VisionBlock.Set");
 export const sendShapeSetNameVisible = sendSimpleShapeOption<boolean>("Shape.Options.NameVisible.Set");
+export const sendShapeSetShowBadge = sendSimpleShapeOption<boolean>("Shape.Options.ShowBadge.Set");
 
 export const sendShapeSetAnnotation = sendSimpleShapeOption<string>("Shape.Options.Annotation.Set");
 export const sendShapeSetName = sendSimpleShapeOption<string>("Shape.Options.Name.Set");