+++ client/src/game/manager.ts
@@ -1,6 +1,5 @@
 import { InvalidationMode, SyncMode } from "@/core/comm/types";
 import { ServerShape } from "@/game/comm/types/shapes";
-import { EventBus } from "@/game/event-bus";
 import { GlobalPoint } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
 import { createShapeFromDict } from "@/game/shapes/utils";
@@ -28,32 +27,6 @@ export class GameManager {
         layer.addShape(sh, SyncMode.NO_SYNC, InvalidationMode.NORMAL);
         layer.invalidate(false);
         return sh;
-    }
-
-    updateShape(data: { shape: ServerShape; redraw: boolean }): void {
-        if (!layerManager.hasLayer(layerManager.getFloor(getFloorId(data.shape.floor))!, data.shape.layer)) {
-            console.log(`Shape with unknown layer ${data.shape.layer} could not be added`);
-            return;
-        }
-        const sh = createShapeFromDict(data.shape);
-        if (sh === undefined) {
-            console.log(`Shape with unknown type ${data.shape.type_} could not be added`);
-            return;
-        }
-        const oldShape = layerManager.UUIDMap.get(data.shape.uuid);
-        if (oldShape === undefined) {
-            console.log(`Shape with unknown id could not be updated`);
-            return;
-        }
-        const redrawInitiative = sh.owners !== oldShape.owners;
-        const shape = Object.assign(oldShape, sh);
-        const alteredVision = shape.checkVisionSources();
-        const alteredMovement = shape.setMovementBlock(shape.movementObstruction, false);
-        shape.setIsToken(shape.isToken, false);
-        if (data.redraw) {
-            shape.updateShapeVision(alteredMovement, alteredVision);
-        }
-        if (redrawInitiative) EventBus.$emit("Initiative.ForceUpdate");
     }
 
     setCenterPosition(position: GlobalPoint): void {