+++ client/src/game/shapes/shape.ts
@@ -43,6 +43,7 @@ import {
     sendShapeUpdateAura,
     sendShapeUpdateTracker,
     sendShapeCreateAura,
+    sendShapeSetShowBadge,
 } from "../api/emits/shape/options";
 import { Floor } from "../layers/floor";
 import { Layer } from "../layers/layer";
@@ -234,7 +235,7 @@ export abstract class Shape {
         this.layer.invalidate(skipLightUpdate);
     }
 
-    checkVisionSources(recalculate = true): boolean {
+    private checkVisionSources(recalculate = true): boolean {
         let alteredVision = false;
         const visionBlockers = getBlockers(TriangulationTarget.VISION, this._floor);
         const obstructionIndex = visionBlockers.indexOf(this.uuid);
@@ -592,7 +593,8 @@ export abstract class Shape {
     setVisionBlock(blocksVision: boolean, sync: boolean, recalculate = true): void {
         if (sync) sendShapeSetBlocksVision({ shape: this.uuid, value: blocksVision });
         this.visionObstruction = blocksVision;
-        this.invalidate(!this.checkVisionSources(recalculate));
+        const alteredVision = this.checkVisionSources(recalculate);
+        if (alteredVision && recalculate) this.invalidate(false);
     }
 
     setMovementBlock(blocksMovement: boolean, sync: boolean, recalculate = true): boolean {
@@ -634,6 +636,12 @@ export abstract class Shape {
     setLocked(isLocked: boolean, sync: boolean): void {
         this.isLocked = isLocked;
         if (sync) sendShapeSetLocked({ shape: this.uuid, value: isLocked });
+    }
+
+    setShowBadge(showBadge: boolean, sync: boolean): void {
+        this.showBadge = showBadge;
+        if (sync) sendShapeSetShowBadge({ shape: this.uuid, value: this.showBadge });
+        this.invalidate(!this.triggersVisionRecalc);
     }
 
     setAnnotation(text: string, sync: boolean): void {