+++ client/src/game/ui/tools/draw.vue
@@ -248,8 +248,8 @@ export default class DrawTool extends Tool implements ToolBasics {
 
             this.shape.addOwner({ user: gameStore.username, access: { edit: true } }, false);
             if (layer.name === "fow" && this.modeSelect === "normal") {
-                this.shape.setVisionBlock(true, false);
-                this.shape.movementObstruction = true;
+                this.shape.setVisionBlock(true, false, false);
+                this.shape.setMovementBlock(true, false, false);
             }
             layer.addShape(this.shape, SyncMode.FULL_SYNC, InvalidationMode.NO);