+++ CHANGELOG.md
@@ -7,6 +7,26 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+-   spawn location tweaks
+    -   Multi spawn locations
+        -   When moving a shape to a new location that has multiple spawn locations, a box will appear to choose the desired spawn zone
+    -   Removal of spawn locations is now possible
+
+### Fixed
+
+-   Run docker container as non-root
+
+# [0.22.2] - 2020-08-28
+
+### Fixed
+
+-   Drawing on FOW layer blocking UI
+-   Badge toggle not working properly
+-   Group leader not properly set serverside on paste
+-   Server error on shape paste due to aura type change
+
 ## [0.22.1] - 2020-08-27
 
 ### Changed